Move object-related files to `(object)`  route group
