+++ apps/researcher/src/app/[locale]/(objects)/heritage-object-card.tsx
